l_sep

l_sep calculates the large frequency separation

Usage

l_sep(model.df, L = NULL, n_inf = NULL, n_sup = NULL, vector = FALSE)

Arguments

model.df

DataFrame

L

Scalar. Mode degree for which the large separation is calculated

n_inf

Scalar. Minimum radial order (n) considered. If NULL (default) minimum is calculated using tostada::nrange function

n_sup

Scalar. Maximum radial order (n) considered. If NULL (default) maximum is calculated using tostada::nrange function

vector

Logical. Indicates wether user want a single value or the list of spacings for the indicated L.

Value

Large separation calculated from frequencies for the chosen L value. If L is not provided then l_sep returns the median value from all the Dnu(L) When n range is not specified, l_sep uses tostada::n_range function to obtain the minimum and maximum n (common to all L values). Otherwise the Dnu value is calculated within the range (n_min, n_max). When vector = T l_sep gives a vector with

If no L specified: Dnu(L=0), Dnu(L=1), ...(Dnu(L=Lmax)

If L specified: dnu_n2-n1, dnu_n3-n2, dnu_n4-n3,...dnu_n-(n-1)

Details

This function takes one argument, a dataframe with the synthetic oscillation spectra read from read_graco, read_adipls, read_gyre, or read_losc, and calculates the large spacing defined as the vector dnu = nu_b - nu_a where nu_b = nu_n,l; nu_a = nu_n-1,l Single value result is calculated as ls = median(dnu) When all L are considered then ls = median(median(dnu,L=0), median(dnu,L=1), ...)

See also

Examples

if (FALSE) {
   my_ls <- l_sep(df_oscillations, L=0, n_inf = 2, n_sup =10) # This
   #calculates the median of the ls for radial modes
   }
if (FALSE) {
  my_ls <- l_sep(df_oscillations, L=0, vector = T) # This
   #calculates the ls for radial modes, and gives a vector with values.
  }