read_cestam

read_cestam reads the content of stellar structure files form CESTAM code, and save it to a dataframe.

Usage

read_cestam(strFile)

Arguments

file

ASCII file with a stellar structure compliant with stellar oscillations computation with extension .osc or .OSC.

Value

A dataframe with the stellar structure, in which global parameters -common values for the star - are saved as attributes, and variables - physical quantities that vary with the stellar depth- are saved as columns.

Details

This function takes one argument: the stellar structure file .osc from CESTAM code. It should also compatible with last versions (deprecated) of CESAM2K models as well as MESA (OSC-type) models. The input file can be local or web address (http).

Metadata

name	name of the model file

code	generator code name

version	version of the code

date	date of the computation

time	time of the computation

Global variables

rmass	mass, cgs

R	radius, cgs

rphot	photometric radius, cgs

rlum	luminosity, cgs

al	convective alpha (dimensionless)

age	age, Myr

rotini	initial angular rotation (in rad/s)

rlogg	surface gravity, cgs, log

teff	effective temperature (temp(r=rphot)), K

nshell	number of shells of the model (radial)

Structure variables

r	radius profile, cgs, log

q	mass profile, cgs, log

temp	temperature profile, K

p	pressure profile, cgs

rho	density profile rho, cgs

gamma	gamma1 parameter, cgs

aast	A constant (BV)

rr	radius normalized to total radius

c1	solar dynamical time

cs2	sound speed squared, cgs

vg	solar dynamical time

rot	rotation profile, rad/s

See also

Examples

mystruc <- read_cestam("/my/path/file.osc")
#> Warning: cannot open file '/my/path/file.osc': No such file or directory
#> Error in file(description = strFile, open = "r"): cannot open the connection

mystruc_from_web <- read_cestam("http://my-path-to-file/myoscfile.osc")
#> Warning: URL 'http://my-path-to-file/myoscfile.osc': status was 'Couldn't resolve host name'
#> Error in file(description = strFile, open = "r"): cannot open the connection to 'http://my-path-to-file/myoscfile.osc'

if (FALSE) {
  mystruc <- read_cestam("/my/path/file.osc")
}