readobs_corot

readobs_corot reads CoRoT FITS files

Usage

readobs_corot(corotfits, hdu = 1)

Arguments

hdu

Numeric, scalar. Data block: 1 (default) or 2,3 (additional data)

filename

String, scalar. Input CoRoT FITS file (.FITS)

Value

A column-named dataframe containing the lightcurve and related data as well as the original header as attributes of the df.

   \strong{AN2_STAR DATA (former sismo field)}

    \it{First data block (hdu = 1)}

    DATETT: (double) date of measurement in Terrestrial Time; terrestrial
    time, Julian date
    RAWFLUX: (float) Integrated red flux; e- per 32s (*)
    RAWFLUXDEV: (float) Standard deviation of the 16 exposures of 32 s in
    the green channel added for the 512 sampling; e- per 32s

. RAWSTATUS: (int) BG: (float) Background flux already subtracted; e- per 32s Notes(*): 32-s or 512-s sampling

    \it{Second data block (hdu = 2)}

    DATEBARTT: (double ) date of the measurement in the solar barycentric
    reference frame; Solar barycentric Terrestrial Time, Julian day
    FLUXBAR: (float) White flux, after the gap filling correction;
    e- per 32s
    FLUXDEVBAR: (float)
    STATUSBAR: (int) Flag of the status

    \it{Second data block (hdu = 3)}

    DATEBARREGTT: (double ) date of the measurement in the solar barycentric
    reference frame; Solar barycentric Terrestrial Time, Julian day
    FLUXBARREG: (float) White flux, after the gap filling correction;
    e- per 32s
    FLUXDEVBARREG: (float)
    STATUSBARREG: (int) Flag of the status

    \strong{EN2_STAR_CH DATA}

    \it{First data block (hdu = 1)}
    DATE: (string) calendar date, yyyy-mm-ddThh:mm:ss
    DATETT: (double) date of measurement in Terrestrial Time; terrestrial
    time, Julian date
    DATEBARTT: (double ) date of the measurement in the solar barycentric
    reference frame; Solar barycentric Terrestrial Time, Julian day
    STATUS: (int) Flag of the status
    REDFLUX: (float) Integrated red flux; e- per 32s (*)
    REDFLUXDEV: (float) Standard deviation of the 16 exposures of 32 s in
    the green channel added for the 512 sampling; e- per 32s
    GREENFLUX: (float) Integrated green flux; e- per 32s (*)
    GREENFLUXDEV: (float) Standard deviation of the 16 exposures of 32 s in
    the green channel added for the 512 sampling; e- per 32s
    BLUEFLUX: (float) Integrated blue flux; e- per 32s (*)
    BLUEFLUXDEV: (float) Standard deviation of the 16 exposures of 32 s in
    the green channel added for the 512 sampling; e- per 32s
    WHITEFLUX: (float) Integrated white flux; e- per 32s
    BG: (float) Background flux already subtracted; e- per 32s
    Notes(*): 32-s or 512-s sampling

    \it{Second data block (hdu = 2)}

    DATETT: (double) date of measurement in Terrestrial Time; terrestrial
    time, Julian date
    DATEBARTT: (double ) date of the measurement in the solar barycentric
    reference frame; Solar barycentric Terrestrial Time, Julian day
    WHITEFLUXFIL: (float) White flux, after the gap filling correction;
    e- per 32s
    STATUSFIL: (int) Flag of the status
    TEXP: (int) Exposure Time; 32s or 512s

    Header data is stored as attributes. To see all
    attributes use:
           attributes(yourobject)
    To select one particular attribute use
           attr(yourobject, "name-of-attribute")

Details

This function reads CoRoT mission FITS files and stores the result in an R dataframe. All the information has been extracted from the publication CoRoT Legacy Book.

Examples

myLC_exo <- readobs_corot('/myhome/EN2_STAR_CHR_corotid_tini-tfin.fits')
#> Warning: cannot open file '/myhome/EN2_STAR_CHR_corotid_tini-tfin.fits': No such file or directory
#> Error in file(file, "rb"): cannot open the connection
myLC_sis <- readobs_corot('/myhome/AN2_STAR_corotid_tini-tfin.fits')
#> Warning: cannot open file '/myhome/AN2_STAR_corotid_tini-tfin.fits': No such file or directory
#> Error in file(file, "rb"): cannot open the connection

if (FALSE) {
  readobs_corot('/myhome/EN2_STAR_CHR_corotid_tini-tfin.fits')
}